+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/AbstractArrowMixin.java
@@ -3,6 +3,7 @@ package insane96mcp.iguanatweaksreborn.mixin;
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
+import insane96mcp.iguanatweaksreborn.module.combat.ArrowStats;
 import insane96mcp.iguanatweaksreborn.module.combat.MiscStats;
 import insane96mcp.iguanatweaksreborn.module.experience.enchantments.EnchantmentsFeature;
 import insane96mcp.iguanatweaksreborn.module.misc.tweaks.Tweaks;
@@ -43,14 +44,14 @@ public abstract class AbstractArrowMixin extends Projectile {
 
     @Redirect(method = "onHitEntity", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/projectile/AbstractArrow;isCritArrow()Z", ordinal = 0))
     private boolean onCritArrowCheck(AbstractArrow arrow) {
-        if (Feature.isEnabled(MiscStats.class) && MiscStats.disableCritArrowsBonusDamage)
+        if (Feature.isEnabled(MiscStats.class) && ArrowStats.disableCritArrowsBonusDamage)
             return false;
         return this.isCritArrow();
     }
 
     @ModifyExpressionValue(method = "tick", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/projectile/AbstractArrow;isCritArrow()Z", ordinal = 0))
     private boolean onCritArrowCheckParticles(boolean original) {
-        if (!Feature.isEnabled(MiscStats.class) || !MiscStats.disableCritArrowsBonusDamage)
+        if (!Feature.isEnabled(MiscStats.class) || !ArrowStats.disableCritArrowsBonusDamage)
             return original;
         return false;
     }
@@ -65,19 +66,19 @@ public abstract class AbstractArrowMixin extends Projectile {
     private double onSetArrowDamage(AbstractArrow instance, Operation<Double> original) {
         if (instance.getOwner() instanceof Mob)
             return original.call(instance);
-        return original.call(instance) * MiscStats.arrowsDamageMultiplier;
+        return original.call(instance) * ArrowStats.damageMultiplier;
     }
 
     @ModifyArg(method = "onHitEntity", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/Entity;hurt(Lnet/minecraft/world/damagesource/DamageSource;F)Z"))
     private float onHurtDamage(float damage) {
-        if (!MiscStats.decimalArrowsDamage)
+        if (!ArrowStats.floatPointDamage)
             return damage;
         double l = this.getDeltaMovement().length();
-        double damageMultiplier = MiscStats.arrowsDamageMultiplier;
+        double damageMultiplier = ArrowStats.damageMultiplier;
         if (this.getOwner() instanceof Mob)
             damageMultiplier = 1f;
         float newDamage = (float) Mth.clamp(l * this.baseDamage * damageMultiplier, 0.0D, Integer.MAX_VALUE);
-        if (this.isCritArrow() && !MiscStats.disableCritArrowsBonusDamage) {
+        if (this.isCritArrow() && !ArrowStats.disableCritArrowsBonusDamage) {
             newDamage += this.random.nextFloat() * (newDamage / 2 + 2);
         }
         return newDamage;